+++ examples/custom_output/custom_output.rs
@@ -9,7 +9,7 @@ use pipeline::{create_pipeline, Vertex};
 use std::time::{Duration, Instant, SystemTime};
 use utils::WgpuUtils;
 use wgpu::util::DeviceExt;
-use wgpu_text::section::{BuiltInLineBreaker, Layout, Section, Text};
+use wgpu_text::glyph_brush::{BuiltInLineBreaker, Layout, Section, Text};
 use wgpu_text::BrushBuilder;
 use winit::{
     event::{ElementState, KeyboardInput, VirtualKeyCode, WindowEvent},