+++ examples/simple.rs
@@ -3,7 +3,7 @@ mod utils;
 
 use std::time::{Duration, Instant, SystemTime};
 use utils::WgpuUtils;
-use wgpu_text::section::{
+use wgpu_text::glyph_brush::{
     BuiltInLineBreaker, Layout, OwnedText, Section, Text, VerticalAlign,
 };
 use wgpu_text::BrushBuilder;