+++ src/lib.rs
@@ -21,34 +21,7 @@ mod error;
 mod pipeline;
 
 pub use brush::{BrushBuilder, TextBrush};
-
-/// Contains all needed objects for inserting, styling and iterating text and glyphs.
-/// Directly taken from glyph_brush.
-///
-/// Look into [`glyph_brush_layout docs`](https://docs.rs/glyph_brush_layout/latest/glyph_brush_layout/#enums)
-/// for the accurate, detailed documentation.
-///
-/// If anything is missing, open an issue on GitHub, and I'll review it.
-pub mod section {
-    #[doc(hidden)]
-    pub use glyph_brush::{
-        BuiltInLineBreaker, Color, FontId, GlyphCruncher, HorizontalAlign, Layout,
-        LineBreak, OwnedSection, OwnedText, Section, SectionGlyphIter, SectionText, Text,
-        VerticalAlign,
-    };
-}
-
-/// Contains all needed objects for font and glyph management.
-/// Directly taken from glyph_brush.
-///
-/// Look into [`glyph_brush_font docs`](https://docs.rs/glyph_brush/latest/glyph_brush/ab_glyph/index.html)
-/// for the accurate, detailed documentation.
-///
-/// If anything is missing, open an issue on GitHub, and I'll review it.
-pub mod font {
-    #[doc(hidden)]
-    pub use glyph_brush::ab_glyph::{Font, FontArc, FontRef, InvalidFont, ScaleFont};
-}
+pub use glyph_brush;
 
 /// Represents a two-dimensional array matrix with 4x4 dimensions.
 pub type Matrix = [[f32; 4]; 4];